// src/types/database.d.ts
export interface o_novel {
 'chapter'?: string | null;
 'chapterData'?: string | null;
 'chapterIndex'?: number | null;
 'createTime'?: number | null;
 'errorReason'?: string | null;
 'event'?: string | null;
 'eventState'?: number | null;
 'id'?: number;
 'projectId'?: number | null;
 'reel'?: string | null;
}